Refatorar método de cálculo de valor líquido
